requirements.txt
streamlit==1.13.0
tensorflow-cpu
Pillow==8.2.0
SQLAlchemy==1.4.35